# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DimensionScore": {
        "properties": {
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "name",
          "score",
          "feedback"
        ],
        "title": "DimensionScore",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobMatchHistoryItem": {
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "job_title": {
            "title": "Job Title",
            "type": "string"
          },
          "match_id": {
            "title": "Match Id",
            "type": "integer"
          },
          "match_score": {
            "title": "Match Score",
            "type": "number"
          }
        },
        "required": [
          "match_id",
          "job_title",
          "company",
          "match_score",
          "created_at"
        ],
        "title": "JobMatchHistoryItem",
        "type": "object"
      },
      "JobMatchHistoryResponse": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/JobMatchHistoryItem"
            },
            "title": "Matches",
            "type": "array"
          }
        },
        "required": [
          "matches"
        ],
        "title": "JobMatchHistoryResponse",
        "type": "object"
      },
      "JobMatchRequest": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "job_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Job Description"
          },
          "job_title": {
            "title": "Job Title",
            "type": "string"
          },
          "mode": {
            "default": "basic",
            "enum": [
              "basic",
              "enhanced"
            ],
            "title": "Mode",
            "type": "string"
          },
          "resume_id": {
            "title": "Resume Id",
            "type": "integer"
          }
        },
        "required": [
          "resume_id",
          "job_title",
          "job_description"
        ],
        "title": "JobMatchRequest",
        "type": "object"
      },
      "JobMatchResponse": {
        "properties": {
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/DimensionScore"
            },
            "title": "Dimensions",
            "type": "array"
          },
          "fit_summary": {
            "title": "Fit Summary",
            "type": "string"
          },
          "full_matches": {
            "items": {
              "type": "string"
            },
            "title": "Full Matches",
            "type": "array"
          },
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "improvement_tips": {
            "items": {
              "type": "string"
            },
            "title": "Improvement Tips",
            "type": "array"
          },
          "match_id": {
            "title": "Match Id",
            "type": "integer"
          },
          "match_score": {
            "title": "Match Score",
            "type": "number"
          },
          "mode": {
            "anyOf": [
              {
                "enum": [
                  "basic",
                  "enhanced"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "partial_matches": {
            "items": {
              "$ref": "#/components/schemas/PartialMatch"
            },
            "title": "Partial Matches",
            "type": "array"
          },
          "provenance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          },
          "required_skills": {
            "items": {
              "type": "string"
            },
            "title": "Required Skills",
            "type": "array"
          },
          "scoring_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Version"
          },
          "skill_verification_rate": {
            "title": "Skill Verification Rate",
            "type": "integer"
          },
          "true_gaps": {
            "items": {
              "type": "string"
            },
            "title": "True Gaps",
            "type": "array"
          },
          "uncertainties": {
            "items": {
              "type": "string"
            },
            "title": "Uncertainties",
            "type": "array"
          }
        },
        "required": [
          "match_id",
          "match_score",
          "grade",
          "required_skills",
          "full_matches",
          "partial_matches",
          "true_gaps",
          "skill_verification_rate",
          "dimensions",
          "fit_summary",
          "improvement_tips"
        ],
        "title": "JobMatchResponse",
        "type": "object"
      },
...
      "PartialMatch": {
        "properties": {
          "coverage": {
            "title": "Coverage",
            "type": "integer"
          },
          "skill": {
            "title": "Skill",
            "type": "string"
          },
          "via": {
            "title": "Via",
            "type": "string"
          }
        },
        "required": [
          "skill",
          "coverage",
          "via"
        ],
        "title": "PartialMatch",
        "type": "object"
      },
...
      "ResumeTailorRequest": {
        "properties": {
          "template_type": {
            "default": "ats",
            "description": "ats, executive, technical, or creative",
            "title": "Template Type",
            "type": "string"
          }
        },
        "title": "ResumeTailorRequest",
        "type": "object"
      },
      "ResumeTailorResponse": {
        "properties": {
          "pdf_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Base64"
          },
          "tailored_resume_markdown": {
            "title": "Tailored Resume Markdown",
            "type": "string"
          }
        },
        "required": [
          "tailored_resume_markdown"
        ],
        "title": "ResumeTailorResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "HireWiz API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/match": {
      "post": {
        "operationId": "match_job_api_jobs_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobMatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobMatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Match Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/match/{match_id}/tailor": {
      "post": {
        "operationId": "tailor_resume_api_jobs_match__match_id__tailor_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeTailorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeTailorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Tailor Resume",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/matches": {
      "get": {
        "operationId": "get_match_history_api_jobs_matches_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobMatchHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Match History",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}